eth/tracers: fix issues with benchmarks #30667
