+++ eth/tracers/tracers_test.go
@@ -98,11 +98,13 @@ func BenchmarkTransactionTrace(b *testing.B) {
 
 	for i := 0; i < b.N; i++ {
 		snap := state.Snapshot()
+		tracer.OnTxStart(evm.GetVMContext(), tx, msg.From)
 		st := core.NewStateTransition(evm, msg, new(core.GasPool).AddGas(tx.Gas()))
-		_, err = st.TransitionDb(common.Address{})
+		res, err := st.TransitionDb(common.Address{})
 		if err != nil {
 			b.Fatal(err)
 		}
+		tracer.OnTxEnd(&types.Receipt{GasUsed: res.UsedGas}, nil)
 		state.RevertToSnapshot(snap)
 		if have, want := len(tracer.StructLogs()), 244752; have != want {
 			b.Fatalf("trace wrong, want %d steps, have %d", want, have)
